requirements.txt
absl-py==2.3.0
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
antlr4-python3-runtime==4.11.0
attrs==25.3.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
DataProperty==1.1.0
datasets==3.6.0
dill==0.3.8
evaluate==0.4.3
filelock==3.18.0
fire==0.7.0
frozenlist==1.6.0
fsspec==2025.3.0
hf-xet==1.1.2
huggingface-hub==0.32.3
human-eval==1.0.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonlines==4.0.0
latex2sympy2_extended==1.10.1
lm_eval==0.4.8
lxml==5.4.0
MarkupSafe==3.0.2
math-verify==0.7.0
mbstrdecoder==1.1.4
modelscope==1.26.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.2.3
pathvalidate==3.2.3
peft==0.15.2
portalocker==3.1.1
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pybind11==2.13.6
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlitedict==2.1.0
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
termcolor==3.1.0
thefuzz==0.22.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.49.0
triton==2.1.0
typepy==1.3.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
word2number==1.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
